generic flipcard component
